added user actions

+++ src/resolvers/mutation.js
@@ -1,4 +1,5 @@
 const bcrypt = require('bcrypt');
+const mongoose = require('mongoose');
 const jwt = require('jsonwebtoken');
 const {
     AuthenticationError,
@@ -11,23 +12,57 @@ const gravatar = require('../util/gravatar');
 
 
 module.exports = {
-    newNote: async (parent, args, { models }) => {
+    // add the users context
+    newNote: async (parent, args, { models, user }) => {
+        // if there is no user on the context, throw an authentication error
+        if (!user) {
+            throw new AuthenticationError('You must be signed in to create a note');
+        }
+
         return await models.Note.create({
             content: args.content,
-            author: 'Jonathan Reeves'
+            // reference the author's mongo id
+            author: mongoose.Types.ObjectId(user.id)
         });
     },
 
-    deleteNote: async (parent, { id }, { models }) => {
+    deleteNote: async (parent, { id }, { models, user }) => {
+        // if not a user, throw an Authentication Error
+        if (!user) {
+            throw new AuthenticationError('You must be signed in to delete a note');
+        }
+
+        // find the note
+        const note = await models.Note.findById(id);
+        // if the note owner and current user don't match, throw a forbidden error
+        if (note && String(note.auther) !== user.id) {
+            throw new ForbiddenError("You don't have permissions to delete the note");
+        }
+
         try {
+            // if everything checks out, remove the note
             await models.Note.findOneAndRemove({ _id: id });
             return true;
         } catch (err) {
+            // if there's an error along the way, return false
             return false;
         }
     },
 
-    updateNote: async (parent, { content, id }, { models }) => {
+    updateNote: async (parent, { content, id }, { models, user }) => {
+        // if not a user, throw an Authentication Error
+        if (!user) {
+            throw new AuthenticationError('You must be signed in to update a note');
+        }
+
+        // find the note
+        const note = await models.Note.findById(id);
+        // if the note owner and current user don't match, throw a forbidden error
+        if (note && String(note.author) !== user.id) {
+            throw new ForbiddenError("You don't have permissions to update the note");
+        }
+        
+        // Update the note in the db and return the updated note
         return await models.Note.findOneAndUpdate(
             {
                 _id: id,
@@ -90,5 +125,53 @@ module.exports = {
 
         // create and return the json web token
         return jwt.sign({ id: user._id }, process.env.JWT_SECRET);
-    }
+    },
+
+    toggleFavorite: async (parent, { id }, { models, user }) => {
+        // if no user context is passed, throw auth error
+        if (!user) {
+            throw new AuthenticationError();
+        }
+
+        // check to see if the user has already favorited the note
+        let noteCheck = await models.Note.findById(id);
+        const hasUser = noteCheck.favoritedBy.indexOf(user.id);
+
+        // if the user exists in the list
+        // pull them from the list and reduce the favirteCount by 1
+        if (hasUser >= 0) {
+            return await models.Note.findByIdAndUpdate(
+                id,
+                {
+                    $pull: {
+                        favoritedBy: mongoose.Types.ObjectId(user.id)
+                    },
+                    $inc: {
+                        favoriteCount: -1
+                    }
+                },
+                {
+                    // Set new to true to return the updated doc
+                    new: true
+                }
+            );
+        } else {
+            // if the user doesn't exist in the list
+            // add them to the list and increment the favoriteCount by 1
+            return await models.Note.findByIdAndUpdate(
+                id,
+                {
+                    $push: {
+                        favoritedBy: mongoose.Types.ObjectId(user.id)
+                    },
+                    $inc: {
+                        favoriteCount: 1
+                    }
+                },
+                {
+                    new: true
+                }
+            );
+        }
+    },
 }
